disable evm debug/tracing
