+++ primitives/evm/rpc/debug/src/lib.rs
@@ -1,28 +1,43 @@
-// Copyright 2019-2021 Bholdus Inc.
-// This file is part of Bholdus.
+// Copyright 2019-2022 PureStake Inc.
+// This file is part of Moonbeam.
 
-// Bholdus is free software: you can redistribute it and/or modify
+// Moonbeam is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
 
-// Bholdus is distributed in the hope that it will be useful,
+// Moonbeam is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
 // You should have received a copy of the GNU General Public License
-// along with Bholdus.  If not, see <http://www.gnu.org/licenses/>.
+// along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
 use codec::{Decode, Encode};
-use ethereum::TransactionV2 as Transaction;
+use ethereum::{TransactionV0 as LegacyTransaction, TransactionV2 as Transaction};
 use ethereum_types::H256;
 use sp_std::vec::Vec;
 
 sp_api::decl_runtime_apis! {
+	// Api version is virtually 4.
+	//
+	// We realized that even using runtime overrides, using the ApiExt interface reads the api
+	// versions from the state runtime, meaning we cannot just reset the versioning as we see fit.
+	//
+	// In order to be able to use ApiExt as part of the RPC handler logic we need to be always
+	// above the version that exists on chain for this Api, even if this Api is only meant
+	// to be used overridden.
+	#[api_version(4)]
 	pub trait DebugRuntimeApi {
+		#[changed_in(4)]
+		fn trace_transaction(
+			extrinsics: Vec<Block::Extrinsic>,
+			transaction: &LegacyTransaction,
+		) -> Result<(), sp_runtime::DispatchError>;
+
 		fn trace_transaction(
 			extrinsics: Vec<Block::Extrinsic>,
 			transaction: &Transaction,
