+++ runtime/ulas/src/lib.rs
@@ -692,86 +692,86 @@ impl_runtime_apis! {
 		}
 	}
 
-	impl bholdus_evm_rpc_primitives_debug::DebugRuntimeApi<Block> for Runtime {
-		fn trace_transaction(
-			extrinsics: Vec<<Block as BlockT>::Extrinsic>,
-			traced_transaction: &EthereumTransaction,
-		) -> Result<
-			(),
-			sp_runtime::DispatchError,
-		> {
-			#[cfg(feature = "evm-tracing")]
-			{
-				use bholdus_evm_tracer::tracer::EvmTracer;
-				// Apply the a subset of extrinsics: all the substrate-specific or ethereum
-				// transactions that preceded the requested transaction.
-				for ext in extrinsics.into_iter() {
-					let _ = match &ext.0.function {
-						Call::Ethereum(transact { transaction }) => {
-							if transaction == traced_transaction {
-								EvmTracer::new().trace(|| Executive::apply_extrinsic(ext));
-								return Ok(());
-							} else {
-								Executive::apply_extrinsic(ext)
-							}
-						}
-						_ => Executive::apply_extrinsic(ext),
-					};
-				}
+	// impl bholdus_evm_rpc_primitives_debug::DebugRuntimeApi<Block> for Runtime {
+	// 	fn trace_transaction(
+	// 		extrinsics: Vec<<Block as BlockT>::Extrinsic>,
+	// 		traced_transaction: &EthereumTransaction,
+	// 	) -> Result<
+	// 		(),
+	// 		sp_runtime::DispatchError,
+	// 	> {
+	// 		#[cfg(feature = "evm-tracing")]
+	// 		{
+	// 			use bholdus_evm_tracer::tracer::EvmTracer;
+	// 			// Apply the a subset of extrinsics: all the substrate-specific or ethereum
+	// 			// transactions that preceded the requested transaction.
+	// 			for ext in extrinsics.into_iter() {
+	// 				let _ = match &ext.0.function {
+	// 					Call::Ethereum(transact { transaction }) => {
+	// 						if transaction == traced_transaction {
+	// 							EvmTracer::new().trace(|| Executive::apply_extrinsic(ext));
+	// 							return Ok(());
+	// 						} else {
+	// 							Executive::apply_extrinsic(ext)
+	// 						}
+	// 					}
+	// 					_ => Executive::apply_extrinsic(ext),
+	// 				};
+	// 			}
 
-				Err(sp_runtime::DispatchError::Other(
-					"Failed to find Ethereum transaction among the extrinsics.",
-				))
-			}
-			#[cfg(not(feature = "evm-tracing"))]
-			Err(sp_runtime::DispatchError::Other(
-				"Missing `evm-tracing` compile time feature flag.",
-			))
-		}
+	// 			Err(sp_runtime::DispatchError::Other(
+	// 				"Failed to find Ethereum transaction among the extrinsics.",
+	// 			))
+	// 		}
+	// 		#[cfg(not(feature = "evm-tracing"))]
+	// 		Err(sp_runtime::DispatchError::Other(
+	// 			"Missing `evm-tracing` compile time feature flag.",
+	// 		))
+	// 	}
 
-		fn trace_block(
-			extrinsics: Vec<<Block as BlockT>::Extrinsic>,
-			known_transactions: Vec<H256>,
-		) -> Result<
-			(),
-			sp_runtime::DispatchError,
-		> {
-			#[cfg(feature = "evm-tracing")]
-			{
-				use bholdus_evm_tracer::tracer::EvmTracer;
-				use sha3::{Digest, Keccak256};
+	// 	fn trace_block(
+	// 		extrinsics: Vec<<Block as BlockT>::Extrinsic>,
+	// 		known_transactions: Vec<H256>,
+	// 	) -> Result<
+	// 		(),
+	// 		sp_runtime::DispatchError,
+	// 	> {
+	// 		#[cfg(feature = "evm-tracing")]
+	// 		{
+	// 			use bholdus_evm_tracer::tracer::EvmTracer;
+	// 			use sha3::{Digest, Keccak256};
 
-				let mut config = <Runtime as pallet_evm::Config>::config().clone();
-				config.estimate = true;
+	// 			let mut config = <Runtime as pallet_evm::Config>::config().clone();
+	// 			config.estimate = true;
 
-				// Apply all extrinsics. Ethereum extrinsics are traced.
-				for ext in extrinsics.into_iter() {
-					match &ext.0.function {
-						Call::Ethereum(transact { transaction }) => {
-							let eth_extrinsic_hash =
-								H256::from_slice(Keccak256::digest(&rlp::encode(transaction)).as_slice());
-							if known_transactions.contains(&eth_extrinsic_hash) {
-								// Each known extrinsic is a new call stack.
-								EvmTracer::emit_new();
-								EvmTracer::new().trace(|| Executive::apply_extrinsic(ext));
-							} else {
-								let _ = Executive::apply_extrinsic(ext);
-							}
-						}
-						_ => {
-							let _ = Executive::apply_extrinsic(ext);
-						}
-					};
-				}
+	// 			// Apply all extrinsics. Ethereum extrinsics are traced.
+	// 			for ext in extrinsics.into_iter() {
+	// 				match &ext.0.function {
+	// 					Call::Ethereum(transact { transaction }) => {
+	// 						let eth_extrinsic_hash =
+	// 							H256::from_slice(Keccak256::digest(&rlp::encode(transaction)).as_slice());
+	// 						if known_transactions.contains(&eth_extrinsic_hash) {
+	// 							// Each known extrinsic is a new call stack.
+	// 							EvmTracer::emit_new();
+	// 							EvmTracer::new().trace(|| Executive::apply_extrinsic(ext));
+	// 						} else {
+	// 							let _ = Executive::apply_extrinsic(ext);
+	// 						}
+	// 					}
+	// 					_ => {
+	// 						let _ = Executive::apply_extrinsic(ext);
+	// 					}
+	// 				};
+	// 			}
 
-				Ok(())
-			}
-			#[cfg(not(feature = "evm-tracing"))]
-			Err(sp_runtime::DispatchError::Other(
-				"Missing `evm-tracing` compile time feature flag.",
-			))
-		}
-	}
+	// 			Ok(())
+	// 		}
+	// 		#[cfg(not(feature = "evm-tracing"))]
+	// 		Err(sp_runtime::DispatchError::Other(
+	// 			"Missing `evm-tracing` compile time feature flag.",
+	// 		))
+	// 	}
+	// }
 
 	#[cfg(feature = "try-runtime")]
 	impl frame_try_runtime::TryRuntime<Block> for Runtime {
